requirements.txt
python-dotenv
telethon
google-auth
google-api-python-client
